+++ audio.cpp
@@ -12,9 +12,7 @@
 #include "apcKey25.h"
 #include <circle/logger.h>
 #include <circle/timer.h>
-#ifdef LOOPER_LIVE_PITCH
-	#include "patches/RubberBandWrapper.h"
-#endif
+#include "patches/RubberBandWrapper.h"
 
 #define log_name "audio"
 
@@ -90,9 +88,7 @@
 	AudioOutputOTG otgOut;
 #endif
 
-#ifdef LOOPER_LIVE_PITCH
-	RubberBandWrapper *pLivePitchWrapper = 0;
-#endif
+RubberBandWrapper *pLivePitchWrapper = 0;
 
 loopMachine *pTheLoopMachine = 0;
 publicLoopMachine *pTheLooper = 0;
@@ -111,16 +107,11 @@ void setup()
 	pTheLooper = (publicLoopMachine *) pTheLoopMachine;
 
 	debug_blink(1);
-#ifdef LOOPER_LIVE_PITCH
 	pLivePitchWrapper = new RubberBandWrapper(AUDIO_SAMPLE_RATE, LOOPER_NUM_CHANNELS);
-	new AudioConnection(input,				0,  *pLivePitchWrapper,	0);
-	new AudioConnection(input,				1,  *pLivePitchWrapper,	1);
-	new AudioConnection(*pLivePitchWrapper,	0,  *pTheLooper,		0);
-	new AudioConnection(*pLivePitchWrapper,	1,  *pTheLooper,		1);
-#else
-	new AudioConnection(input,			0,  *pTheLooper,	0);
-	new AudioConnection(input,			1,  *pTheLooper,	1);
-#endif
+	new AudioConnection(input,              0,  *pLivePitchWrapper, 0);
+	new AudioConnection(input,              1,  *pLivePitchWrapper, 1);
+	new AudioConnection(*pLivePitchWrapper, 0,  *pTheLooper,        0);
+	new AudioConnection(*pLivePitchWrapper, 1,  *pTheLooper,        1);
 	new AudioConnection(*pTheLooper,	0,  output,			0);
 	new AudioConnection(*pTheLooper,	1,  output,			1);
 
@@ -168,10 +159,8 @@ void setup()
 
 void loop()
 {
-#ifdef LOOPER_LIVE_PITCH
-	if (pLivePitchWrapper && pTheAPC && pTheAPC->m_transposeLocked)
+	if (pLivePitchWrapper)
 		pLivePitchWrapper->updateRatios();
-#endif
 	if (pTheLooper) {
 		logString_t *msg;
 		while ((msg = pTheLooper->getNextLogString()) != nullptr) {